requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pandas>=2.0.0
boto3>=1.28.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.4.0
